requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
python-dotenv==1.0.1
gunicorn==23.0.0
